+++ src/middlewares/passport/index.ts
@@ -1,54 +1,18 @@
 import User from "../../models/user-model";
-import passport from "passport";
-// import kakaoStrategy from "passport-kakao"
-const KakaoStrategy = require("passport-kakao").Strategy;
-import { kakao } from "../../config/constants";
-import { UserParams } from "../../common/type";
+import KakaoStratege from "./kakao-strategy";
 
-const KakaoModule = (app: any) => {
-  app.use(passport.initialize());
-
-  passport.use(
-    new KakaoStrategy(
-      {
-        clientID: kakao.kakaoId,
-        callbackURL: kakao.kakaoUrl,
-      },
-      async (
-        accessToken: string,
-        refreshToken: string,
-        profile: any,
-        done: any
-      ) => {
-        try {
-          const existUser: UserParams | null = await User.findOne({
-            email: profile._json.kakao_account.email,
-          });
-
-          if (existUser) {
-            done(null, existUser);
-          }
-
-          const newUser: UserParams = await User.create({
-            email: profile._json.kakao_account.email,
-            nickname: profile._json.properties.nickname,
-            profileImg: profile._json.properties.profile_image,
-          });
-
-          done(null, newUser);
-        } catch (error) {
-          console.error(error);
-          done(error);
-        }
-      }
-    )
-  );
-  passport.serializeUser((user, done) => {
+const kakaoPassport = (passport: any) => {
+  passport.serializeUser((user: any, done: any) => {
     done(null, user);
   });
-  passport.deserializeUser((user, done) => {
-    done(null, user as any);
+
+  passport.deserializeUser((user: any, done: any) => {
+    User.findOne({ _id: user.userId })
+      .then((user) => done(null, user))
+      .catch((err) => done(err));
   });
+
+  KakaoStratege(passport);
 };
 
-export default KakaoModule;
+export default kakaoPassport;
